+++ src/editor/core/command/Command.ts
@@ -10,6 +10,7 @@ export class Command {
   public executeBackspace: CommandAdapt['backspace']
   public executeSetRange: CommandAdapt['setRange']
   public executeForceUpdate: CommandAdapt['forceUpdate']
+  public executeBlur: CommandAdapt['blur']
   public executeUndo: CommandAdapt['undo']
   public executeRedo: CommandAdapt['redo']
   public executePainter: CommandAdapt['painter']
@@ -83,6 +84,7 @@ export class Command {
   public executeDeleteGroup: CommandAdapt['deleteGroup']
   public executeLocationGroup: CommandAdapt['locationGroup']
   public executeSetControlValue: CommandAdapt['setControlValue']
+  public executeSetControlExtension: CommandAdapt['setControlExtension']
   public getCatalog: CommandAdapt['getCatalog']
   public getImage: CommandAdapt['getImage']
   public getOptions: CommandAdapt['getOptions']
@@ -99,6 +101,7 @@ export class Command {
   public getLocale: CommandAdapt['getLocale']
   public getGroupIds: CommandAdapt['getGroupIds']
   public getControlValue: CommandAdapt['getControlValue']
+  public getContainer: CommandAdapt['getContainer']
 
   constructor(adapt: CommandAdapt) {
     // 全局命令
@@ -110,6 +113,7 @@ export class Command {
     this.executeBackspace = adapt.backspace.bind(adapt)
     this.executeSetRange = adapt.setRange.bind(adapt)
     this.executeForceUpdate = adapt.forceUpdate.bind(adapt)
+    this.executeBlur = adapt.blur.bind(adapt)
     // 撤销、重做、格式刷、清除格式
     this.executeUndo = adapt.undo.bind(adapt)
     this.executeRedo = adapt.redo.bind(adapt)
@@ -205,8 +209,10 @@ export class Command {
     this.getSearchNavigateInfo = adapt.getSearchNavigateInfo.bind(adapt)
     this.getLocale = adapt.getLocale.bind(adapt)
     this.getGroupIds = adapt.getGroupIds.bind(adapt)
+    this.getContainer = adapt.getContainer.bind(adapt)
     // 控件
     this.executeSetControlValue = adapt.setControlValue.bind(adapt)
+    this.executeSetControlExtension = adapt.setControlExtension.bind(adapt)
     this.getControlValue = adapt.getControlValue.bind(adapt)
   }
 }