+++ src/editor/core/command/CommandAdapt.ts
@@ -21,7 +21,8 @@ import { DeepRequired } from '../../interface/Common'
 import {
   IGetControlValueOption,
   IGetControlValueResult,
-  ISetControlOption
+  ISetControlExtensionOption,
+  ISetControlValueOption
 } from '../../interface/Control'
 import {
   IAppendElementListOption,
@@ -40,7 +41,7 @@ import {
 } from '../../interface/Editor'
 import { IElement, IElementStyle } from '../../interface/Element'
 import { IMargin } from '../../interface/Margin'
-import { RangeContext } from '../../interface/Range'
+import { RangeContext, RangeRect } from '../../interface/Range'
 import { IColgroup } from '../../interface/table/Colgroup'
 import { ITd } from '../../interface/table/Td'
 import { ITr } from '../../interface/table/Tr'
@@ -169,6 +170,11 @@ export class CommandAdapt {
       isSubmitHistory,
       isSetCursor: false
     })
+  }
+
+  public blur() {
+    this.range.clearRange()
+    this.draw.getCursor().recoveryCursor()
   }
 
   public undo() {
@@ -360,8 +366,6 @@ export class CommandAdapt {
   public superscript() {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
-    const activeControl = this.control.getActiveControl()
-    if (activeControl) return
     const selection = this.range.getSelectionElementList()
     if (!selection) return
     const superscriptIndex = selection.findIndex(
@@ -391,8 +395,6 @@ export class CommandAdapt {
   public subscript() {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
-    const activeControl = this.control.getActiveControl()
-    if (activeControl) return
     const selection = this.range.getSelectionElementList()
     if (!selection) return
     const subscriptIndex = selection.findIndex(
@@ -1808,12 +1810,52 @@ export class CommandAdapt {
     const positionList = this.position.getPositionList()
     const startPageNo = positionList[startIndex].pageNo
     const endPageNo = positionList[endIndex].pageNo
+    // 坐标信息（相对编辑器书写区）
+    const rangeRects: RangeRect[] = []
+    const selectionPositionList = this.position.getSelectionPositionList()
+    if (selectionPositionList) {
+      const height = this.draw.getOriginalHeight()
+      const pageGap = this.draw.getOriginalPageGap()
+      // 起始信息及x坐标
+      let currentRowNo: number | null = null
+      let currentX = 0
+      let rangeRect: RangeRect | null = null
+      for (let p = 0; p < selectionPositionList.length; p++) {
+        const {
+          rowNo,
+          pageNo,
+          coordinate: { leftTop, rightTop },
+          lineHeight
+        } = selectionPositionList[p]
+        // 起始行变化追加选区信息
+        if (currentRowNo === null || currentRowNo !== rowNo) {
+          if (rangeRect) {
+            rangeRects.push(rangeRect)
+          }
+          rangeRect = {
+            x: leftTop[0],
+            y: leftTop[1] + pageNo * (height + pageGap),
+            width: rightTop[0] - leftTop[0],
+            height: lineHeight
+          }
+          currentRowNo = rowNo
+          currentX = leftTop[0]
+        } else {
+          rangeRect!.width = rightTop[0] - currentX
+        }
+        // 最后一个元素结束追加选区信息
+        if (p === selectionPositionList.length - 1 && rangeRect) {
+          rangeRects.push(rangeRect)
+        }
+      }
+    }
     return deepClone({
       isCollapsed,
       startElement,
       endElement,
       startPageNo,
-      endPageNo
+      endPageNo,
+      rangeRects
     })
   }
 
@@ -1874,11 +1916,12 @@ export class CommandAdapt {
     if (!payload.length) return
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
+    const cloneElementList = deepClone(payload)
     // 格式化上下文信息
     const { startIndex } = this.range.getRange()
     const elementList = this.draw.getElementList()
-    formatElementContext(elementList, payload, startIndex)
-    this.draw.insertElementList(payload)
+    formatElementContext(elementList, cloneElementList, startIndex)
+    this.draw.insertElementList(cloneElementList)
   }
 
   public appendElementList(
@@ -1888,7 +1931,7 @@ export class CommandAdapt {
     if (!elementList.length) return
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
-    this.draw.appendElementList(elementList, options)
+    this.draw.appendElementList(deepClone(elementList), options)
   }
 
   public setValue(payload: Partial<IEditorData>) {
@@ -1900,10 +1943,11 @@ export class CommandAdapt {
     if (startIndex !== endIndex) return
     const elementList = this.draw.getElementList()
     const element = elementList[startIndex]
-    if (element.type !== ElementType.CONTROL) return
+    if (!element.controlId) return
     // 删除控件
     const control = this.draw.getControl()
     const newIndex = control.removeControl(startIndex)
+    if (newIndex === null) return
     // 重新渲染
     this.range.setRange(newIndex, newIndex)
     this.draw.render({
@@ -2037,9 +2081,19 @@ export class CommandAdapt {
     return this.draw.getControl().getValueByConceptId(payload)
   }
 
-  public setControlValue(payload: ISetControlOption) {
+  public setControlValue(payload: ISetControlValueOption) {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
     this.draw.getControl().setValueByConceptId(payload)
+  }
+
+  public setControlExtension(payload: ISetControlExtensionOption) {
+    const isReadonly = this.draw.isReadonly()
+    if (isReadonly) return
+    this.draw.getControl().setExtensionByConceptId(payload)
+  }
+
+  public getContainer(): HTMLDivElement {
+    return this.draw.getContainer()
   }
 }