+++ src/editor/core/range/RangeManager.ts
@@ -275,7 +275,7 @@ export class RangeManager {
     if (~startIndex && ~endIndex) {
       const elementList = this.draw.getElementList()
       const element = elementList[startIndex]
-      if (element?.type === ElementType.CONTROL) {
+      if (element?.controlId) {
         control.initControl()
         return
       }
@@ -416,6 +416,7 @@ export class RangeManager {
     const elementList = this.draw.getElementList()
     const range = this.getRange()
     const { startIndex, endIndex } = range
+    if (!~startIndex && !~endIndex) return
     const startElement = elementList[startIndex]
     const endElement = elementList[endIndex]
     if (startIndex === endIndex) {